+++ pkg/bitboard/bitscan_test.go
@@ -9,7 +9,7 @@ import (
 
 func TestLeastSignificantOneBit(t *testing.T) {
 	assert.Panics(t, func() { LeastSignificantOneBit(Empty) })
-	for i := 0; i < 64; i++ {
+	for i := uint8(0); i < 64; i++ {
 		t.Run(fmt.Sprintf("LSB %v", i), func(t *testing.T) {
 			if got := LeastSignificantOneBit(1 << i); got != i {
 				t.Errorf("LeastSignificantOneBit() = %v, want %v", got, i)
@@ -20,7 +20,7 @@ func TestLeastSignificantOneBit(t *testing.T) {
 
 func TestMostSignificantOneBit(t *testing.T) {
 	assert.Panics(t, func() { MostSignificantOneBit(Empty) })
-	for i := 0; i < 64; i++ {
+	for i := uint8(0); i < 64; i++ {
 		t.Run(fmt.Sprintf("LSB %v", i), func(t *testing.T) {
 			if got := MostSignificantOneBit(1 << (63 - i)); got != i {
 				t.Errorf("LeastSignificantOneBit() = %v, want %v", got, i)
