+++ pkg/bitboard/squares.go
@@ -2,7 +2,7 @@ package bitboard
 
 import "github.com/shaardie/clemens/pkg/types"
 
-func BitBySquares(squares ...int) Bitboard {
+func BitBySquares(squares ...uint8) Bitboard {
 	b := Empty
 	for _, s := range squares {
 		b |= One << s
@@ -10,14 +10,14 @@ func BitBySquares(squares ...int) Bitboard {
 	return b
 }
 
-func RankMaskOfSquare(square int) Bitboard {
+func RankMaskOfSquare(square uint8) Bitboard {
 	return RankMask1 << (8 * types.RankOfSquare(square))
 }
 
-func FileMaskOfSquare(square int) Bitboard {
+func FileMaskOfSquare(square uint8) Bitboard {
 	return FileMaskA << types.FileOfSquare(square)
 }
 
-func BitboardFromRankAndFile(rank int, file int) Bitboard {
+func BitboardFromRankAndFile(rank uint8, file uint8) Bitboard {
 	return BitBySquares(types.SquareFromRankAndFile(rank, file))
 }
